+++ src/main/java/com/openclassrooms/rental/repository/MessageRepository.java
@@ -3,5 +3,5 @@ package com.openclassrooms.rental.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.openclassrooms.rental.model.MessageEntity;
 
-public interface MessageRepository extends CrudRepository<MessageEntity, Integer> {
+public interface MessageRepository extends CrudRepository<MessageEntity, Long> {
 }
